Rename to .tsx files
