+++ site/src/components/sdkTable.tsx
@@ -74,7 +74,6 @@ export default class Table extends React.Component {
             </tbody>
             </table>
           </div>
-
         );
     }
 }
